DOC: Rename folders in economic_curves
    * Remove &, - from folder names
    * Black_Sholes.py -> Black_Scholes.py
