+++ lifelib/libraries/economic_curves/smith_wilson/LICENSE
