+++ lifelib/libraries/economic_curves/smith_wilson/Smith&Wilson_example.xlsx
