+++ lifelib/libraries/economic_curves/stationary_bootstrap/LICENSE
